Move validateNpmName into validate-project-path.ts
